+++ src/components/Button/Button.stories.tsx
@@ -1,15 +1,12 @@
 import { VStack } from "@chakra-ui/react";
 import type { Meta, StoryObj } from "@storybook/react";
 
-import Button, {
-  buttonElementTypeArray,
-  buttonSizesArray,
-  buttonTypesArray,
-} from "./Button";
+import Button, { buttonElementTypeArray, buttonTypesArray } from "./Button";
 import ButtonGroup from "../ButtonGroup/ButtonGroup";
 import Icon from "../Icons/Icon";
 import { iconNamesArray } from "../Icons/iconVariables";
 import Table from "../Table/Table";
+import { sizesArray } from "../../theme/sharedTypes";
 
 const meta: Meta<typeof Button> = {
   title: "Components/Form Elements/Button",
@@ -43,7 +40,7 @@ const meta: Meta<typeof Button> = {
     onClick: { control: false },
     size: {
       control: { type: "radio" },
-      options: buttonSizesArray,
+      options: sizesArray,
       table: { defaultValue: { summary: "medium" } },
     },
     type: {
