+++ src/components/Image/Image.tsx
@@ -23,6 +23,7 @@ export const imageRatiosArray = [
   "threeByTwo",
   "twoByOne",
 ] as const;
+
 export const imageSizesArray = [
   "default",
   "xxxsmall",
