+++ src/components/StatusBadge/StatusBadge.tsx
@@ -1,5 +1,6 @@
 import { Box, chakra, ChakraComponent, useStyleConfig } from "@chakra-ui/react";
 import React, { forwardRef } from "react";
+import { messageTypeArray } from "../../theme/sharedTypes";
 
 export const statusBadgeFontSizeArray = [
   "desktop.body.body1",
@@ -7,15 +8,7 @@ export const statusBadgeFontSizeArray = [
   "desktop.caption",
 ] as const;
 
-export const statusBadgeTypeArray = [
-  "informative",
-  "negative",
-  "neutral",
-  "positive",
-  "recommendation",
-  "warning",
-] as const;
-export type StatusBadgeTypes = typeof statusBadgeTypeArray[number];
+export type StatusBadgeTypes = typeof messageTypeArray[number];
 export interface StatusBadgeProps {
   /** Additional class for the component */
   className?: string;
