+++ src/components/VideoPlayer/VideoPlayer.tsx
@@ -14,6 +14,7 @@ export const VideoPlayerAspectRatiosArray = [
   "sixteenByNine",
   "square",
 ] as const;
+
 export const VideoPlayerTypesArray = ["vimeo", "youtube"] as const;
 export type VideoPlayerAspectRatios =
   typeof VideoPlayerAspectRatiosArray[number];
